requirements.txt
#Django==3.0.5
django-admin-interface==0.19.1
django-colorfield==0.7.1
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-js-asset==2.0.0
djangorestframework==3.13.1
#psycopg2==2.8.6
#django-heroku==0.3.1
#gunicorn==20.1.0
idna==3.3
#mysqlclient==2.1.0
Pillow==9.1.1
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
